+++ lib/api_client.dart
@@ -21,19 +21,8 @@ class ApiClient {
 
   final String basePath;
 
-  var _client = Client();
-
-  /// Returns the current HTTP [Client] instance to use in this class.
-  ///
-  /// The return value is guaranteed to never be null.
-  Client get client => _client;
-
-  /// Requests to use a new HTTP [Client] in this class.
-  ///
-  /// If the [newClient] is null, an [ArgumentError] is thrown.
-  set client(Client newClient) {
-    _client = newClient;
-  }
+  /// The HTTP [Client] instance to use in this class.
+  var client = Client();
 
   final _defaultHeaderMap = <String, String>{};
   final _authentications = <String, Authentication>{};
@@ -99,7 +88,7 @@ class ApiClient {
               onError: (Object error, StackTrace trace) => request.sink.close(),
               cancelOnError: true,
             );
-        final response = await _client.send(request);
+        final response = await client.send(request);
         return Response.fromStream(response);
       }
 
@@ -109,7 +98,7 @@ class ApiClient {
         request.files.addAll(body.files);
         request.headers.addAll(body.headers);
         request.headers.addAll(headerParams);
-        final response = await _client.send(request);
+        final response = await client.send(request);
         return Response.fromStream(response);
       }
 
@@ -120,36 +109,36 @@ class ApiClient {
 
       switch (method) {
         case 'POST':
-          return await _client.post(
+          return await client.post(
             uri,
             headers: nullableHeaderParams,
             body: msgBody,
           );
         case 'PUT':
-          return await _client.put(
+          return await client.put(
             uri,
             headers: nullableHeaderParams,
             body: msgBody,
           );
         case 'DELETE':
-          return await _client.delete(
+          return await client.delete(
             uri,
             headers: nullableHeaderParams,
             body: msgBody,
           );
         case 'PATCH':
-          return await _client.patch(
+          return await client.patch(
             uri,
             headers: nullableHeaderParams,
             body: msgBody,
           );
         case 'HEAD':
-          return await _client.head(
+          return await client.head(
             uri,
             headers: nullableHeaderParams,
           );
         case 'GET':
-          return await _client.get(
+          return await client.get(
             uri,
             headers: nullableHeaderParams,
           );